code revamp, fixed search suggestions api
