+++ src/App.js
@@ -1,36 +1,23 @@
 import './App.css';
-import Body from './components/Body';
-import Head from './components/Head';
-import WatchPage from './components/WatchPage';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import Layout from './Layout';
+import Navbar from './components/Navbar';
+import WatchPage from './pages/WatchPage';
 import { Provider } from 'react-redux';
 import store from './utils/store';
-import MainContainer from './components/MainContainer';
+import HomePage from './pages/HomePage';
+import { Route, Routes } from 'react-router-dom';
 
-const appRouter = createBrowserRouter([
-  {
-    path: "/",
-    element: <Body/>,
-    children: [
-      {
-        path: "/",
-        element: <MainContainer />,
-      },
-      {
-        path: "watch",
-        element: <WatchPage />,
-      },
-    ],
-  },
-]);
 function App() {
   return (
     <Provider store={store}>
-    <div>
-      <Head />
-      <RouterProvider router={appRouter} />
-    </div>
-  </Provider>
+          <Navbar />
+          <Routes>
+            <Route path="/" element={<Layout />}>
+              <Route index element={<HomePage />} />
+              <Route path='watch' element={<WatchPage />} />
+            </Route>
+          </Routes>
+    </Provider>
   );
 }
