+++ src/components/CommentsContainer.jsx
@@ -1,114 +1,14 @@
 import React from "react";
-
-const commentsData = [
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [],
-  },
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [
-      {
-        name: "John Doe",
-        text: "Lorem ipsum dolor sit amet, consectetur adip",
-        replies: [],
-      },
-      {
-        name: "John Doe",
-        text: "Lorem ipsum dolor sit amet, consectetur adip",
-        replies: [
-          {
-            name: "John Doe",
-            text: "Lorem ipsum dolor sit amet, consectetur adip",
-            replies: [
-              {
-                name: "John Doe",
-                text: "Lorem ipsum dolor sit amet, consectetur adip",
-                replies: [
-                  {
-                    name: "John Doe",
-                    text: "Lorem ipsum dolor sit amet, consectetur adip",
-                    replies: [
-                      {
-                        name: "John Doe",
-                        text: "Lorem ipsum dolor sit amet, consectetur adip",
-                        replies: [],
-                      },
-                    ],
-                  },
-                  {
-                    name: "John Doe",
-                    text: "Lorem ipsum dolor sit amet, consectetur adip",
-                    replies: [],
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [],
-  },
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [],
-  },
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [],
-  },
-  {
-    name: "John Doe",
-    text: "Lorem ipsum dolor sit amet, consectetur adip",
-    replies: [],
-  },
-];
-
-const Comment = ({ data }) => {
-  const { name, text, replies } = data;
-  return (
-    <div className="flex shadow-sm bg-gray-100 p-2 rounded-lg my-2">
-      <img
-        className="w-8 h-8 md:w-12 md:h-12"
-        alt="user"
-        src="https://www.iconpacks.net/icons/2/free-user-icon-3296-thumb.png"
-      />
-      <div className="px-1 md:px-3">
-        <p className="font-bold">{name}</p>
-        <p className="text-sm md:text-base">{text}</p>
-      </div>
-    </div>
-  );
-};
-
-const CommentsList = ({ comments }) => {
-  
-  return comments.map((comment, index) => (
-    <div key={index}>
-      <Comment data={comment} />
-      <div className="pl-5 border-l border-l-black ml-5">
-        <CommentsList comments={comment.replies} />
-      </div>
-    </div>
-  ));
-};
+import { commentsData } from "../utils/constants";
+import CommentsList from "./CommentList";
 
 const CommentsContainer = () => {
-  return (
-    <div>
-      <h1 className="text-2xl font-bold">Comments: </h1>
-      <CommentsList comments={commentsData} />
-    </div>
-  );
+	return (
+		<div>
+			<h1 className="text-2xl font-bold">Comments: </h1>
+			<CommentsList comments={commentsData} />
+		</div>
+	);
 };
 
 export default CommentsContainer;
