+++ src/pages/HomePage.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import FilterList from './FilterList'
-import VideoList from './VideoList'
+import FilterList from '../components/FilterList'
+import VideoList from '../components/VideoList'
 
-const MainContainer = () => {
+const HomePage = () => {
   return (
     <div className='flex flex-col gap-4 w-11/12 md:w-10/12 mx-auto mb-8'>
         <FilterList />
@@ -11,4 +11,4 @@ const MainContainer = () => {
   )
 }
 
-export default MainContainer
+export default HomePage
