+++ src/utils/chatSlice.js
@@ -1,16 +1,16 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 const chatSlice = createSlice({
-  name: "chat",
-  initialState: {
-    messages: [],
-  },
-  reducers: {
-    addMessage: (state, action) => {
-      if (state.messages.length === 25) state.messages.shift(); // implement a queue 
-      state.messages.push(action.payload);
-    },
-  },
+	name: "chat",
+	initialState: {
+		messages: [],
+	},
+	reducers: {
+		addMessage: (state, action) => {
+			if (state.messages.length === 25) state.messages.shift();
+			state.messages.push(action.payload);
+		},
+	},
 });
 
 export const { addMessage } = chatSlice.actions;
