+++ src/utils/constants.js
@@ -1,0 +1,72 @@
+export const commentsData = [
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [],
+    },
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [
+        {
+          name: "John Doe",
+          text: "Lorem ipsum dolor sit amet, consectetur adip",
+          replies: [],
+        },
+        {
+          name: "John Doe",
+          text: "Lorem ipsum dolor sit amet, consectetur adip",
+          replies: [
+            {
+              name: "John Doe",
+              text: "Lorem ipsum dolor sit amet, consectetur adip",
+              replies: [
+                {
+                  name: "John Doe",
+                  text: "Lorem ipsum dolor sit amet, consectetur adip",
+                  replies: [
+                    {
+                      name: "John Doe",
+                      text: "Lorem ipsum dolor sit amet, consectetur adip",
+                      replies: [
+                        {
+                          name: "John Doe",
+                          text: "Lorem ipsum dolor sit amet, consectetur adip",
+                          replies: [],
+                        },
+                      ],
+                    },
+                    {
+                      name: "John Doe",
+                      text: "Lorem ipsum dolor sit amet, consectetur adip",
+                      replies: [],
+                    },
+                  ],
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [],
+    },
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [],
+    },
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [],
+    },
+    {
+      name: "John Doe",
+      text: "Lorem ipsum dolor sit amet, consectetur adip",
+      replies: [],
+    },
+  ];
